Securidad en E y N

+++ controllers/NoticiasController.php
@@ -17,39 +17,55 @@ class NoticiasController extends Noticia
     function storeNew()
     {
         $fecha = new DateTime();
-        // funcion img
         $directorio = "assets/uploud/news/";
         $archivo= $directorio.basename($fecha->getTimeStamp().$_FILES["new_img"]["name"]);
         //info de ext(jpg,png,etc)
         $tipoArchivo =strtolower(pathinfo($archivo,PATHINFO_EXTENSION));
         //verifica que el archivo tenga dimensiones(w,h) 
         $DimensionesImg =getimagesize($_FILES['new_img']['tmp_name']);
+
+        //POST
+        $tituloNoticia = Security::htmlChars($_POST['titulo_noticia']);
+        $descripcionNoticia = Security::htmlChars($_POST['descripcion_noticia']);
+        $fechaPublicacion = Security::verificateDate($_POST['fecha_noticia']);
+        $newUser = Security::verificateInt($_POST['fk_usuario']);
+
         
         if($DimensionesImg == true)
         {
             $tamañoImg = $_FILES['new_img']["size"];
             if($tamañoImg > 2000000)
             {
-                echo "El archivo tiene que ser menor a 2mb";
+                // echo "El archivo tiene que ser menor a 2mb";
+                echo json_encode(['error'=>'El<2mb']);
             }
             else{
                 if($tipoArchivo == "jpg" || $tipoArchivo == "png" || $tipoArchivo == "jpeg" )
                 {
-                    move_uploaded_file($_FILES["new_img"]["tmp_name"],$archivo);
-                    $tituloNoticia = $_POST['titulo_noticia'];
-                    $descripcionNoticia = $_POST['descripcion_noticia'];
-                    $fechaPublicacion = $_POST['fecha_noticia'];
-                    $imgNew = $archivo;
-                    $newUser = $_POST['fk_usuario'];
+                    
+                    if($tituloNoticia && $descripcionNoticia && $fechaPublicacion && $newUser)
+                    {
+                        move_uploaded_file($_FILES["new_img"]["tmp_name"],$archivo);
+                        $imgNew = $archivo;
+                        parent::storeAddNew($tituloNoticia,$descripcionNoticia,$fechaPublicacion,$imgNew,$newUser);
+                        echo json_encode(['ok'=>'CreadaNoticia']);
+
+                    }else{
+                        echo json_encode(['error'=>'errorCrearNoticia']);
+                    }
             
-                    parent::storeAddNew($tituloNoticia,$descripcionNoticia,$fechaPublicacion,$imgNew,$newUser);
                 }
                 else{
-                    echo "la extension del archivo no es valida";
+                    echo json_encode(['error'=>'noImgExtension']);
+                    header('location=?c=All&m=index');
+                    //no es una img 
                 }
             }
         }else{
-            echo "el documento no es una img";
+            echo json_encode(['error'=>'noImg']);
+            // echo "el documento no es una img";
+            header('location=?c=All&m=index');
+
         }
     
 
@@ -58,60 +74,72 @@ class NoticiasController extends Noticia
 
     public function updateNews()
     {
-        $idNoticia =$_POST['update_id_noticia'];
-        $tituloNoticia = $_POST['update_titulo_noticia'];
-        $descripcionNoticia = $_POST['update_descripcion_noticia'];
-        $fechaPublicacion = $_POST['update_fecha_noticia'];
-        $idUser = $_POST['update_fk_usuario'];
-        //para la ruta de la imagen 
-        $rutaImg=parent::showImg($idNoticia)->imagen_noticia;
-        
-        $imgUpdate = '';
-        //? si no se actualiza una img tomara el valor que tenia anteriormente y si actualiza se borrara la anterior
-        if(empty($_POST['update_new_img']))
+        $idNoticia =Security::verificateInt($_POST['update_id_noticia']);
+        $tituloNoticia = Security::htmlChars($_POST['update_titulo_noticia']);
+        $descripcionNoticia =Security::htmlChars($_POST['update_descripcion_noticia']);
+        $fechaPublicacion = Security::verificateDate($_POST['update_fecha_noticia']);
+        $idUser = Security::verificateInt($_POST['update_fk_usuario']);
+        if($idNoticia && $fechaPublicacion && $idUser && $descripcionNoticia && $tituloNoticia)
         {
-            //? Update Img
-            $fecha = new DateTime();
-            // funcion img
-            $directorio = "assets/uploud/news/";
-            $archivo= $directorio.basename($fecha->getTimeStamp().$_FILES["update_new_img"]["name"]);
-            //info de ext(jpg,png,etc)
-            $tipoArchivo =strtolower(pathinfo($archivo,PATHINFO_EXTENSION));
-            //verifica que el archivo tenga dimensiones(w,h) 
-            $DimensionesImg =getimagesize($_FILES['update_new_img']['tmp_name']);
+            //para la ruta de la imagen 
+            $rutaImg=parent::showImg($idNoticia)->imagen_noticia;
 
-            if($DimensionesImg == true)
+            $imgUpdate = '';
+            //? si no se actualiza una img tomara el valor que tenia anteriormente y si actualiza se borrara la anterior
+            if(empty($_POST['update_new_img']))
             {
-                $tamañoImg = $_FILES['update_new_img']["size"];
-                if($tamañoImg > 2000000)
+                //? Update Img
+                $fecha = new DateTime();
+                // funcion img
+                $directorio = "assets/uploud/news/";
+                $archivo= $directorio.basename($fecha->getTimeStamp().$_FILES["update_new_img"]["name"]);
+                //info de ext(jpg,png,etc)
+                $tipoArchivo =strtolower(pathinfo($archivo,PATHINFO_EXTENSION));
+                //verifica que el archivo tenga dimensiones(w,h) 
+                $DimensionesImg =getimagesize($_FILES['update_new_img']['tmp_name']);
+
+                if($DimensionesImg == true)
                 {
-                    echo "El archivo tiene que ser menor a 2mb";
-                }
-                else{
-                    if($tipoArchivo == "jpg" || $tipoArchivo == "png" || $tipoArchivo == "jpeg" )
+                    $tamañoImg = $_FILES['update_new_img']["size"];
+                    if($tamañoImg > 2000000)
                     {
-                        //Borra la imagen que tenia antes
-                        unlink($rutaImg);
-                        move_uploaded_file($_FILES["update_new_img"]["tmp_name"],$archivo);
-                        $imgUpdate=$archivo;
+                        // echo "El archivo tiene que ser menor a 2mb";
+                        echo json_encode(['error'=>"El archivo tiene que ser menor a 2mb"]);
                     }
                     else{
-                        echo "la extension del archivo no es valida";
+                        if($tipoArchivo == "jpg" || $tipoArchivo == "png" || $tipoArchivo == "jpeg" )
+                        {
+                            //Borra la imagen que tenia antes
+                            unlink($rutaImg);
+                            move_uploaded_file($_FILES["update_new_img"]["tmp_name"],$archivo);
+                            $imgUpdate=$archivo;
+                        }
+                        else{
+                            // echo "la extension del archivo no es valida";
+                            echo json_encode(['error'=>'a extension del archivo no es valida']);
+                        }
                     }
+                }else{
+                    // echo "el documento no es una img";
+                     echo json_encode(['error'=> 'el documento no es una img']);
+
                 }
             }else{
-                echo "el documento no es una img";
+                $imgUpdate = $rutaImg;
             }
+
+            parent::updateNew($tituloNoticia,$descripcionNoticia,$fechaPublicacion,$imgUpdate,$idUser,$idNoticia);
+            echo json_encode(['ok'=> 'ActualizacionNoticia']);
         }else{
-            $imgUpdate = $rutaImg;
+            echo json_encode(['error'=> 'FalloActualizacionNoticia']);
+        
         }
 
-        parent::updateNew($tituloNoticia,$descripcionNoticia,$fechaPublicacion,$imgUpdate,$idUser,$idNoticia);
+
     }
 
     public function destroyNew()
     {
-
         $id = $_REQUEST['id'];
         //funcion para mostrar la ruta de la img por id
         $rutaImg = parent::showImg($id);
